Remove redundant rotate/scale/translate methods
