+++ Storyteller/Storyteller/Model/Layer/Layer Component/CompositeComponent.swift
@@ -16,18 +16,6 @@ extension CompositeComponent: LayerComponent {
         CompositeComponent(children: children.map({ $0.transformed(using: transform) }))
     }
 
-    func scaled(by scale: CGFloat) -> CompositeComponent {
-        CompositeComponent(children: children.map({ $0.scaled(by: scale) }))
-    }
-
-    func rotated(by rotation: CGFloat) -> CompositeComponent {
-        CompositeComponent(children: children.map({ $0.rotated(by: rotation) }))
-    }
-
-    func translatedBy(x: CGFloat, y: CGFloat) -> CompositeComponent {
-        CompositeComponent(children: children.map({ $0.translatedBy(x: x, y: y) }))
-    }
-
     var canvasSize: CGSize {
         children.first?.canvasSize ?? .zero
     }
