+++ Storyteller/Storyteller/Model/Layer/Layer Component/Transformable.swift
@@ -8,7 +8,4 @@ import CoreGraphics
 
 protocol Transformable {
     func transformed(using transform: CGAffineTransform) -> Self
-    func scaled(by scale: CGFloat) -> Self
-    func rotated(by rotation: CGFloat) -> Self
-    func translatedBy(x: CGFloat, y: CGFloat) -> Self
 }
